categories table

+++ client/src/components/menu/Menu.jsx
@@ -5,6 +5,7 @@ import ellipsis from '../../assets/ellipsis.svg';
 import Edit from '../../assets/Edit.svg';
 import { Link } from "react-router-dom";
 import { deleteRecipe } from "../../services/ApiServices";
+import Swal from 'sweetalert';
 
 const Menu = ({ currentPage, recipeId, recipes, setRecipes }) => {
 
@@ -17,9 +18,24 @@ const Menu = ({ currentPage, recipeId, recipes, setRecipes }) => {
     const handleDelete = async (e, id) => {
         try {
             e.stopPropagation();
-            await deleteRecipe(id);
-            const newRecipeList = recipes.filter(el => el.id !== id);
-            setRecipes(newRecipeList);
+
+            Swal({
+                title: "Eliminar receta",
+                text: "Estás seguro de que deseas eliminar esta receta?",
+                icon: "warning",
+                buttons: true,
+                dangerMode: true,
+            })
+                .then((willDelete) => {
+                    if (willDelete) {
+                        deleteRecipe(id);
+                        const newRecipeList = recipes.filter(el => el.id !== id);
+                        setRecipes(newRecipeList);
+                        Swal("Receta eliminada con éxito", {
+                            icon: "success",
+                        });
+                    }
+                });
         } catch (error) {
             console.error('Error deleting recipe:', error);
         }
@@ -39,8 +55,7 @@ const Menu = ({ currentPage, recipeId, recipes, setRecipes }) => {
                                 <span>Editar Receta</span></Link>
                             </li>
                             <li className="list-style-none">
-                                <button onClick={(e) => handleDelete(e, recipeId)} className='delete-button'><img src={Delete} alt="" /></button>
-                                <span>Eliminar receta</span>
+                                <button onClick={(e) => handleDelete(e, recipeId)} className='delete-button'><img src={Delete} alt="" /><span>Eliminar receta</span></button>
                             </li>
                         </>
                     )}
